refactor: move typings to @ast-decorators/utils
