+++ packages/core/package.json
@@ -11,7 +11,6 @@
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "dependencies": {
-    "@ast-decorators/typings": "^0.1.0",
     "@ast-decorators/utils": "^0.1.0",
     "minimatch": "^3.0.4"
   },
