+++ packages/core/tests/fixtures/plugin-options/transformer-params/transformer.ts
@@ -1,5 +1,5 @@
-import {ASTClassDecorator, PrivacyType} from '@ast-decorators/typings';
-import {NodePath} from '@babel/traverse';
+import {ASTClassDecorator, PrivacyType} from '@ast-decorators/utils/lib/common';
+import {NodePath} from '@babel/core';
 import {ClassBody, privateName, StringLiteral} from '@babel/types';
 
 export type TransformerOptions = {privacy?: PrivacyType};
@@ -8,12 +8,15 @@ export type BabelObject = {
   types: typeof import('@babel/types');
 };
 
-export default ({types: t}: BabelObject, {privacy}: TransformerOptions = {}) => [
+export default (
+  {types: t}: BabelObject,
+  {privacy}: TransformerOptions = {},
+) => [
   [
     (
       name: NodePath<StringLiteral>,
       value: NodePath<StringLiteral>,
-    ): ASTClassDecorator => (klass) => {
+    ): ASTClassDecorator => klass => {
       const classBody = klass.get('body') as NodePath<ClassBody>;
       const id = classBody.scope.generateUidIdentifier(name.node.value);
       const node =
