+++ packages/transform-accessor/src/getter.ts
@@ -1,7 +1,7 @@
 import {
   ASTClassMemberCallableDecorator,
   ClassMemberMethod,
-} from '@ast-decorators/typings';
+} from '@ast-decorators/utils/lib/common';
 import {NodePath} from '@babel/core';
 import {
   blockStatement,
