+++ packages/transform-accessor/src/setter.ts
@@ -1,7 +1,7 @@
 import {
   ASTClassMemberCallableDecorator,
   ClassMemberMethod,
-} from '@ast-decorators/typings';
+} from '@ast-decorators/utils/lib/common';
 import {NodePath} from '@babel/core';
 import {
   assignmentExpression,
@@ -22,8 +22,8 @@ import {
   AccessorInterceptorNode,
   AccessorMethodCreator,
   createAccessorDecorator,
-  ownerNode,
   injectInterceptor,
+  ownerNode,
 } from './utils/misc';
 
 export const createSetterMethod: AccessorMethodCreator = (
