--- packages/typings/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@ast-decorators/typings",
-  "version": "0.1.0",
-  "description": "AST Decorators core plugin",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "scripts": {
-    "typecheck": "tsc -p tsconfig.json --noEmit"
-  },
-  "peerDependencies": {
-    "@babel/types": "^7",
-    "@types/babel__core": "^7"
-  }
-}
