--- packages/typings/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "include": ["lib/**/*"]
-}
