+++ tsconfig.json
@@ -22,14 +22,12 @@
   },
   "include": [
     "packages/**/*",
-    "packages/typings/lib/**/*",
     "utils/**/*",
     "jest.config.js"
   ],
   "exclude": [
     "node_modules/!(@open-wc)",
     "packages/*/tests/fixtures/**/*",
-    "packages/!(typings)/lib/**/*",
     "packages/*/node_modules/**/*",
     "packages/*/lib/**/*"
   ]
